requirements.txt
chainlit
agentic
google-generativeai
python-dotenv
